requirements.txt
Pillow==7.2.0
pkg-resources==0.0.0
pymongo==3.11.0
